- Material::AtomCount() now returns double; - Reimplemented in Geant4GM and RootGM to be consistent with AgddGM, which uses these values as input - In test/agddv6/Materials.agdd: replaced the element mass fractions with atom numbers - In AGDDWriter::WriteMaterial(): write atom count instead of mass fraction

+++ packages/AgddGM/include/AgddGM/materials/Material.h
@@ -1,4 +1,4 @@
-// $Id: Element.h,v 1.1 2007-01-25 21:57:12 bviren Exp $
+// $Id$
 
 // -----------------------------------------------------------------------
 // The AgddGM package of the Virtual Geometry Model
@@ -50,7 +50,7 @@ public:
     int     NofElements() const;
     VGM::IElement*  Element(int iel) const;
     double  MassFraction(int iel) const;
-    int AtomCount(int iel) const;
+    double  AtomCount(int iel) const;
 
 private:
